findbugs
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@1227 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/common/Convenience.java
@@ -173,7 +173,11 @@ public class Convenience {
 	 */
 	private File getToAttach(String dir, File file,
 			BiPredicate<File, File> mediaTester, File toAttach) {
-		for (File member : file.listFiles()) {
+		final File[] listFiles = file.listFiles();
+		if (listFiles == null) {
+			return toAttach;
+		}
+		for (File member : listFiles) {
 			File memberFile = new File(dir, member.getName());
 			memberFile.deleteOnExit();
 			if (memberFile.isFile() && isSupportedMedia(memberFile)
